+++ custom/aaa/reports/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-from custom.reach.reports.reports import DashboardReport
+from custom.aaa.reports.reports import DashboardReport
 
 CUSTOM_REPORTS = (
     ('CUSTOM REPORTS', (
